requirements.txt
flask
python-dotenv
gunicorn
waitress
psycopg2-binary
